Move android related tests files to compiled module
