+++ java-checks-test-sources/src/main/java/android/content/Intent.java
@@ -1,0 +1,4 @@
+package android.content;
+
+public class Intent {
+}
