+++ java-checks-test-sources/src/main/java/android/content/IntentFilter.java
@@ -1,0 +1,4 @@
+package android.content;
+
+public class IntentFilter {
+}
