+++ java-checks-test-sources/src/main/java/checks/security/AndroidBroadcastingCheck.java
@@ -1,14 +1,13 @@
+package checks.security;
+
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.UserHandle;
-import android.support.annotation.RequiresApi;
 
-public class MyIntentBroadcast {
-  @RequiresApi(api = Build.VERSION_CODES.JELLY_BEAN_MR1)
+public class AndroidBroadcastingCheck {
   public void broadcast(Intent intent, Context context, UserHandle user,
                         BroadcastReceiver resultReceiver, Handler scheduler, int initialCode,
                         String initialData, Bundle initialExtras,
