+++ java-checks-test-sources/src/main/java/checks/security/ReceivingIntentsCheck.java
@@ -1,13 +1,12 @@
+package checks.security;
+
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.IntentFilter;
-import android.os.Build;
 import android.os.Handler;
-import android.support.annotation.RequiresApi;
 
-public class MyIntentReceiver {
+public class ReceivingIntentsCheck {
 
-  @RequiresApi(api = Build.VERSION_CODES.O)
   public void register(Context context, BroadcastReceiver receiver,
                        IntentFilter filter,
                        String broadcastPermission,
