+++ java-checks/src/test/java/org/sonar/java/checks/security/AndroidBroadcastingCheckTest.java
@@ -22,19 +22,18 @@ package org.sonar.java.checks.security;
 import org.junit.jupiter.api.Test;
 import org.sonar.java.checks.verifier.CheckVerifier;
 
+import static org.sonar.java.checks.verifier.TestUtils.ANDROID_FAKE_CLASS_PATH;
+import static org.sonar.java.checks.verifier.TestUtils.testSourcesPath;
+
 class AndroidBroadcastingCheckTest {
 
   @Test
   void test() {
     CheckVerifier.newVerifier()
-      .onFile("src/test/files/checks/security/AndroidBroadcasting.java")
+      .onFile(testSourcesPath("checks/security/AndroidBroadcastingCheck.java"))
+      .withClassPath(ANDROID_FAKE_CLASS_PATH)
       .withCheck(new AndroidBroadcastingCheck())
       .verifyIssues();
-    CheckVerifier.newVerifier()
-      .onFile("src/test/files/checks/security/AndroidBroadcasting.java")
-      .withCheck(new AndroidBroadcastingCheck())
-      .withoutSemantic()
-      .verifyNoIssues();
   }
 
 }
