+++ java-checks/src/test/java/org/sonar/java/checks/security/AndroidExternalStorageCheckTest.java
@@ -22,19 +22,18 @@ package org.sonar.java.checks.security;
 import org.junit.jupiter.api.Test;
 import org.sonar.java.checks.verifier.CheckVerifier;
 
+import static org.sonar.java.checks.verifier.TestUtils.ANDROID_FAKE_CLASS_PATH;
+import static org.sonar.java.checks.verifier.TestUtils.testSourcesPath;
+
 class AndroidExternalStorageCheckTest {
 
   @Test
   void test() {
     CheckVerifier.newVerifier()
-      .onFile("src/test/files/checks/security/AndroidExternalStorage.java")
+      .onFile(testSourcesPath("checks/security/AndroidExternalStorage.java"))
       .withCheck(new AndroidExternalStorageCheck())
+      .withClassPath(ANDROID_FAKE_CLASS_PATH)
       .verifyIssues();
-    CheckVerifier.newVerifier()
-      .onFile("src/test/files/checks/security/AndroidExternalStorage.java")
-      .withCheck(new AndroidExternalStorageCheck())
-      .withoutSemantic()
-      .verifyNoIssues();
   }
 
 }
